# Use a parseless phrase db to speed up LM loading