--- a/Source/Data/bin/buildFreq.py
+++ b/Source/Data/bin/buildFreq.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 import sys
 import math